requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.9.3
SQLAlchemy==1.4.31
python-dotenv==0.19.2
gunicorn==20.1.0
Flask-Migrate==3.1.0
flask-swagger-ui==3.36.0
requests==2.26.0
Flask-Limiter==1.3.1
pytest==6.2.5